+++ services/ui-src/src/measures/2024/shared/globalValidations/index.ts
@@ -11,7 +11,7 @@ export * from "./validateAtLeastOneDeviationFieldFilled";
 export * from "shared/globalValidations/validateAtLeastOneRateComplete";
 export * from "shared/globalValidations/validateBothDatesInRange";
 export * from "shared/globalValidations/validateDateRangeRadioButtonCompletion";
-export * from "./validateDualPopInformation";
+export * from "shared/globalValidations/validateDualPopInformation";
 export * from "./validateEqualCategoryDenominators";
 export * from "shared/globalValidations/validateEqualQualifierDenominators";
 export * from "./validateFfsRadioButtonCompletion";
