+++ services/ui-src/src/measures/2024/shared/util/validationsMock.tsx
@@ -5,7 +5,7 @@ import * as validateAtLeastOneDeliverySystem from "shared/globalValidations/vali
 import * as validateAtLeastOneDeviationFieldFilled from "measures/2024/shared/globalValidations/validateAtLeastOneDeviationFieldFilled";
 import * as validateAtLeastOneRateComplete from "shared/globalValidations/validateAtLeastOneRateComplete";
 import * as validateBothDatesInRange from "shared/globalValidations/validateBothDatesInRange";
-import * as validateDualPopInformation from "measures/2024/shared/globalValidations/validateDualPopInformation";
+import * as validateDualPopInformation from "shared/globalValidations/validateDualPopInformation";
 import * as validateEqualCategoryDenominators from "measures/2024/shared/globalValidations/validateEqualCategoryDenominators";
 import * as validateEqualQualifierDenominators from "shared/globalValidations/validateEqualQualifierDenominators";
 import * as validateFfsRadioButtonCompletion from "measures/2024/shared/globalValidations/validateFfsRadioButtonCompletion";
